Fix for Bug#: 3539671
Updated the configureComponent page so the alert messages will be handled
and displayed correctly.  With this fix, the operator will be told if a
field has invalid input and will be informed why the input value was invalid.
This commit will also fix the bug where the save option was not
being processed on the first attempt after an alert message was displayed.

Note: This fix was also committed into the beta2 branch.

Reviewed By: Mike Write

Tests-run: QL, jbi-test


QL Test results:

 pecount-java:

      [java] ************************
      [java] PASSED=   69
      [java] ------------  =========
      [java] FAILED=   0
      [java] ------------  =========
      [java] DID NOT RUN=   0
      [java] ------------  =========
      [java] Total Expected=69
      [java] ************************