IN=107407 - JPQL: fix invalid SQL for simple COUNT DISTINCT query
This integration fixes the generated SQL for a simple JPQL query using
a COUNT DISTINCT on the range variable, e.g
  SELECT COUNT(DISTINCT s) FROM Student s

This fix extends the change to fix issue 2497 "JPQL: fix invalid SQL for COUNT
on joined variable". The old code handled a COUNT aggregate in the SELECT clause
 only if the COUNT argument was a QueryKeyExpression representing a field access.
Now it is able to handle an ExpressionBuilder represnting the COUNT argument.

Issue number: 2725
Reviewed by: Tom Ware, Chris Delahunt
Tests : entity-persistence-tests, local tests