Issue number:   6514847 and several others Obtained from:  Sun's Bug Database Submitted by:   Several Reviewed by:    admin@glassfish
This bug is the root cause of several problems in
startup of various server instances in a cluster
environment.

The problem was because of lack of refreshed config
context in node-agent VM for launching various server
instances. All the problems that required recycling of
node-agent for dynamic reconfiguration and/or server
and cluster startup should be gone with this fix.

Byron contributed to this fix.

PE/EE QL were run successfully. Also, manual testing was done.

Learnt yet another thing: Don't ignore "definite assignment"
and "definite unassignment" of blank final variables.