Issue number:  bug# 6501080 Reviewed by:  Prashanth Abbagani
The monitor command disables the key echoing and listens to keyboard inputs.  The native code that disables key echoing uses a low level call to the terminal device.  If  key "q" or "Q" is entered, the monitor command resets the terminal device and the command exits gracefully.  However, if "Ctrl-C" is entered to kill the process, the terminal will be in a state where keys are not echoing back to the terminal.  By adding the trap in asadmin and set the key echoing will fix this problem.

Tested by executing the "monitor" command and stop the process with "Ctrl-C" and verified that no side effects in the terminal.
Tested this fix on Solaris, Mac and Linux.
On Windows, there are no side effects.  No changes required in asadmin.bat.