Fix for Bug#: 6543949
Fix the bug in the CLI (asadmin) where a failure message was displayed
after a command was completed successfully, but with warnings.

When the user tried to uninstall a JBI shared library and that library was
being used by an installed but currently disabled component.  The resulting
message would warning the user about the dependent component issue, and then
state that the command failed, when it actually should of stated that the
command was successful.

Fix Implemented:
  Very simple fix - Evaluate the returned exception result for a WARNING label
  before assuming it is an error.

Tests-run: jbi-test