Fix for Bug#: 6543151
The current code assumed that an exception result was always a failure.
This logic needed to be changed to evaluate the result message and check
to see if it is a "WARNING" message instead of an error message.

Now for this use case, the user will see a "Warning" alert message,
specifying that what they have done could cause a problem in the future.
Note, if the user tries to delete more then one entry in the table and
one entry causes an error, while the other causes a warning, the user
will see the "Error" alert message dialog containing both the error
message and the warning message.