Fix for:   Issue 2688 - Webstart and JPA - no suitable driver found   Issue 2856 - web start deployment only works if ejbs are looked up throug[h   (unreported missing localization in error dialog box for Java Web Start support)
Problem:
2688 - The appclient script included the TopLinkEssentials and the Derby client library on the system class path via the CLASSPATH environment variable before starting the ACC.  TopLinkEssentials (TLE) uses the class loader that loaded its classes to locate the database driver.  This worked if the app client used Derby, because TLE and the Derby driver were loaded by the system class loader.  But if the user packaged a different driver with the application TLE would be unable to locate it.  This is because the TLE class could search for the driver only using the system class loader because that's the loader that loaded the TLE classes.  The user-provided driver would be accessible to the EJBClassLoader created by the ACC but TLE did not use that EJBClassLoader.

The fix removes the TLE and Derby client library (and also the dbschema.jar) from the system class path and specifies them instead on a system property passed to the ACC.  The fix also modifies the ACC so that it retrieves the value of the system property and uses it to augment the class path for the EJBClassLoader created for use by the app client.  Since TLE is now loaded by that EJBClassLoader that also has access to all the user-provided classes, TLE can find the user-provided driver.


2856 - During Java Web start launches Java Web Start causes the event dispatcher thread (EDT) to be initialized and started before the ACC code runs.  This sets the context class loader for the event dispatcher thread to the current class loader of Java Web Start.  The ACC sets its own context class loader - but (formerly) on the main thread only.  As a result, code that ran in response to a user interface gesture would not use the ACC's class loader and so would not have access to those classes.

The fix explicitly sets the context class loader for the EDT to the EJBClassLoader created during initialization in the case of a Java Web Start launch.  During a traditional appclient script launch, the EDT starts - if ever - when the user code first activates a user interface component and the EDT's class loader is initialized to be the same as the main thread's, which is the desired EJBClassLoader.

(localization in error screen) - The LocalStrings property had been added as part of the original error screen addition but the code used a hard-coded English value rather than retrieving the property value.


Fix:

Reviewed by: Hong


Tests:

PE SQE quick-looks; manual test of secure injected and JNDI look-up from app client to EJB; test of example attached to Issue 2688


Code changes:

appserv-core/src/java/com/sun/enterprise/appclient
  AppClientInfo.java
  AppClientInfoFactory.java
  MainWithModuleSupport.java

appserv-core/src/java/com/sun/enterprise/appclient/jws
  ClassPathManager.java
  ErrorDisplayDialog.java
  JWSACCMain.java

appserv-core/misc/templates
  appclient.bat.template
  appclient.template