Fix for CR #6545440
Fix for the managed target page.
The managed target page returned the user to a blank screen after
the Ok or Cancel button was clicked.  The problem was caused by a
syntax error in the calling jsf file.  An extra } was placed at the
end of the navigate command

Also with this integration, a fix for the 404 error that was being
displayed when the user clicked the cancel button when the canceling
from either of the two install component screens.

Reviewed By: Mike Write

Tests-run: QL, jbi-test