IN=107967 - JPQL: fix invalid error message for embedded with property-based access
Fix a bug in the validation phase of the JPQL query compiler. Navigation through
an embedded inside of a JPQL query fails, if the Embeddable uses property-based
persistence and the underlying field has a different name as the property, e.g.

  SELECT e FROM Employee e WHERE e.country.code = 'de'

  @Embeddable
  public class Country implements Serializable {
    private String countryCode;
    @Basic
    public String getCode() { return countryCode; }
    ...
  }

The current implementation uses a class object to represent a type and uses the
class object to look up a ClassDescriptor in the session. As I understand, the
ClassDescriptor as stored in the session for an embeddable is only a shell and
does not hold all the information. Instead I need to use the ClassDescriptor
from the AggregateMapping of the embedded field. The fix is to use the
ClassDescriptor (instead of the class object) as the type representation in
case of an embedded instance. Then there is no need to lookup the ClassDescriptor
in the session when analyzing a field access of the embedded.

Issue number: 2883
Reviewed by: Tom Ware
Tests : entity-persistence-tests, local tests