Fix for CR #6550741 & CR #6550704
Add the column sort functionality for the three main tables (Shared Libraries,
Components and Deployments).

Add the logic in the code that will disable the Save buttons as well as the
dropdown list on the JBI logger page, if the component does not have an
loggers associative  to it.

Reviewed By: Mike Write

Tests-run: jbi-test