Fix for CR # 6554896
Add logic to the code that will allow addition "glassfish" loggers
to be displayed and managed from with in the JBI component logger
page.  Currently this is only needed for the sun-javaee-engine"
component, but other components can also use this feature by
updating the "AdditonalLogger.xml" file.

Tests-run: QL, jbi-test

Reviewed by: Mike Wright