Fix for Issue 2993 - Deletion of autodeployed directory on Windows causes errors
Problem: With the added support for autodeploying directories (as well as archives), auto-undeploying would require deleting the autodeployed directory.  But this causes two problems - JAR files in the app may stay locked because the containers have not been notified of the undeployemnt yet.  And the deletion of files in the top-level autodeployed directory cause the directory file's timestamp to be updated, which the autodeployer interprets as a redeployment.  The attempted redeployment typically fails because many of the files in the autodeployed directory have been deleted.

Fix: Allow detection of xxx_undeployRequested files in the autodeploy directory.  The creation of such a file is a request for the autodeployer to undeploy the corresponding autodeployed app xxx (archive or directory) followed by the deletion of that file or directory and the deletion of the xxx_undeployRequested file.

Reviewed by: self (Hong is away)


Tests:

PE SQE quick-looks; autodeployment devtests


Code changes:

appserv-core/src/java/com/sun/enterprise/deployment/autodeploy/AutoDeployConstants.java
appserv-core/src/java/com/sun/enterprise/deployment/autodeploy/AutoDeployer.java
appserv-core/src/java/com/sun/enterprise/deployment/autodeploy/AutoDeployedFilesManager.java