Issue number:   3003 Obtained from:  Issue Tracker Submitted by:   Self Reviewed by:    Self
This mostly completes the remaining implementation for the "upgrade" of
a domain that does not support clusters to the one that does.

Tested manually by creating a domain that does not support clusters (aka
dev profile domain) and then upgrading it, restarting it and then creating
node-agent, cluster with two instances against it.

This should turn out to be a useful feature.

No impact on QL as QL is not invoking this path.

Also, decremented the redefinition of a constant.