Issue number:   3003 Obtained from:  IT Submitted by:   Self Reviewed by:    Self
This is more of a cleanup.
Domain creation code was incorrectly replacing SERVER_ID (which is
supposed to be the name of the server) for default-virtual-server
ID. This worked because both have the same values ("server"). But
it is wrong to replace the tokens SERVER_ID with "server" when it
comes to default-virtual-server.

What it boils down to, is that the name of the virtual server (default)
is not to be tokenized. It has to be a fixed string, "server".

This change just replaces "%%%SERVER_ID%%%" with "server".

Ran the EE QL against cluster profile domain. The changes are stereotyped
and are not to be seen by end-users.