Fix for CR # 6560800
The requested glassfish loggers can now be viewed and set from the jbi
logger page.  This was to accommodate jbi components that use glassfish
loggers as well as jbi loggers.  This bug was written against the code
because setting the logger levels were not be done successfully.

Also updated the String.properties file to fix some typos.  Changed
Integtration to Integration in 6 different places.

Tests-run: QL, jbi-test

Reviewed by: Mike Wright