IN=108281 - Fix NPE in runtime code populating the result for an outer fetch join query
Fix a NPE in ForeignReferenceMapping.buildClone thrown in case a JPQL query defines
an outer fetch join for a OneToOne relationship on the non-owning side, e.g.
  SELECT i FROM Item i LEFT JOIN FETCH i.order
With the relationship definition in class Item:
  @OneToOne(mappedBy="item")
  public Order getOrder() { ... }

Method buildClone checks the type of the cloned attribute value in order to call the getValue
method if it is a value holder. This check runs into the NPE, in case the attribute value is
null. I changed the code to call getValue only if the attribute value is not null.

Issue number: 2881
Reviewed by: Tom Ware
Tests : entity-persistence-tests, local tests