Fix for CR #6550737 and #6550739
Add logic that will display an alert warning message when the user tries
to "Disable" or "Enable" a component or SA when the target is down.  Note.
the code is not completely  functional yet, since the  "isTargetUp() common
client method is not available yet..  All the code has been tested using a stub
routine.  Another integration will be performed once the method is available, and
the CR's will then be closed.

Tests-run: QL, jbi-test

Reviewed by: Mike Wright