IN=108432 - Fix invalid SQL generated when mixing inner and outer joins
Fix a problem with the generated SQL for a JPQL query mixing inner and outer
joins when running on derby. The generated SQL places one of the outer join
clauses in the wrong part of the SQL FROM clause.

The fix is to reorder the outer join expressions such that all outer join expressions
having the same source alias are processed together. Then the generated SQL includes
the outer join clause in the correct order.

Issue number: 3041
Reviewed by: Tom Ware
Tests : entity-persistence-tests, local tests