Improve Grizzly integration:
+ Add support for StreamAlgorithm
+ Improve header and body bytes manipulation
+ Allow caching

I've added a couple of TODO: to make sure those issues are addressed. The most important one is we need an http mapper when dispatching (for performance reason) to an Adapter as request in the form of http://...:8080//test will never reach a container because of the double //. The current for loop is also dangerous for scripting attack as it can loop forever.

This work is in preparation of integrating Grizzly 1.5.x