Issue number:   3135 Obtained from:  Issue Tracker Submitted by:   JF Reviewed by:    Self
Making enterprise profile to have -server by
default. This is a recommendation to cover
some low-end machines. Thus, we don't leave it
up to the Java SE 5.0 VM ergonomics to decide
whether a Server VM be started.

This is an enterprise profile only change.

As a result, the server startup might be slower ...

Will make the changes to the BETA-branch shortly ...

Ran Quicklook tests ...