Issue number:   3135, 6567180 Obtained from:  Issue Tracker /Sun's database Submitted by:   JF/Deep Singh Reviewed by:    Self.
This is a small but subtle change on the BETA branch.
I have tested it thoroughly, manually, although this is
mostly reverting back the original changes to web-tier config.

- According to perf team suggestion, we should be running "server"
  VM by default.
- It is observed that it is not always right to have too many
  worker threads. So, the values are changed from "10-10-130" back
  to "2-5-1", where the numbers indicate: initial-count:increment-count:max-count.
  Thus, we create 2 worker threads per listener and increment it by 1 till 5.

As a result, on my sunblade, the server (enterprise profile) starts significantly slower
as it uses server VM. But this is the intent as enterprise domain and servers
are not restarted frequently.

Approved by: Dhiru.