Issue number: 6568497 [BLOCKING] asupgrade is broken Reviewed by:  Hans Hrasna
There was a logic error in the handling of the options.  There was also
an ordering problem between the calls to recover and helpUsage.  The
helpUsage message was eventually calling System.exit which prevented
recover from ever being invoked.  This left things in a corrupt state
that prevented subsequent invocations of asupgrade from working.CVS: ----------------------------------------------------------------------