Fix for Issue #3174
Added code to the InstanceSelectionHandlers.java file that will filter the list of
instances to only contain the instances where the component has been installed on.
Also added logic that will return the warning alert information that will be
displayed when the component is not installed to any target.

Added jsftemplating code to the configureComponent, configureLoggers and
monitorComponent files that will call the new updated handler method
jbiGetInstances.  The files will now display an alert message if the component
has not been installed to any target instance.

Reviewed By: Mike Write

Tests-run: QL, jbi-test