Issue 2859.
During the system app loading, when bean pool is initialized a separate thread
is spawned and that causes a JNDI entry context to reach EJB service group.
Changed the code to avoid the thread spawned from the system app loading to
not block the server startup.

Reviewed by Sankar.
Ran QL, and ondemand devtests.