Fix for Issue 3248 - App client deployed from NetBeans not accessible after server restart
Problem: A stand-alone app client deployed using the NetBeans 5.5.1 plug-in is accessible via Java Web Start immediately after deployment until the server is restarted.  Thereafter the client is no longer JWS-accessible until it is deployed again, and only then until the next restart.

This is actually a more general problem with an app client module deployed using the JSR-88 API (which is what the plug-in uses).  The JSR-88 model is that the deployment client invokes DeploymentManager.distribute to deliver the application to whichever targets might be appropriate, then separately "enables" the module on potentially some subset of those targets using the DeploymentManager.start method.  To follow this model our implementation of the JSR-88 API marks the module as disabled during the distribute step.  Because currently the app module MBean does not support the enabled attribute, the start operation of the JSR-88 API is a no-op for app client modules.  So the app client module is never actually enabled on the appropriate targets.

Fix: The long-term fix, for a variety of reasons, is to add the "enabled" attribute to the app client module MBean and support changes to its value, including config change event notification.  This would require changes to the domain.xml DTD, the app client module serverbean, the ApplicationsConfigMBean (which currently rejects attempts to enable or disable app client modules), and the ChangeStateAction in GlassFish.  And those changes would not by themselves resolve this issue without a change in the plug-in.  (Earlier the plug-in was changed to avoid invoking the start or stop methods on the JSR-88 deployment manager - see issue 641.  Those changes would need to be reverted for the above changes to have any effect.)

Because of the potential risks in changing domain.xml and the other related classes at this stage in GlassFish V2, this check-in provides a short-term fix.  This change makes sure that whenever an application ref is created for an app client module that it is always enabled.

This has not been an issue with other ways of deploying app client modules, because the admin CLI (for example) specifies enabled as true by default so the app client module is enabled where it should be.

Reviewed by: Hong

Tests:

PE SQE quick-looks; deployment devtests; manual tests of the problem scenarios (using both the default server and a separate instance)


Code changes:

appserv-core/src/java/com/sun/enterprise/deployment/phasing/PEDeploymentService.java