Issue number: Sun's database: 6517284 Obtained from: Bugster Submitted by:  Sankar Neelakandan Reviewed by:   Abhijit, Self.
Thanks to Abhijit for his help.

The bug has been fixed such that we create a dummy cluster event
listener so that we actually cover the deficincies in the event
infrastructure, which is sending an event when the target is
"domain". The event is sent to ALL the servers in the domain,
affecting everyone.

Ran QL. Did manual testing.