Fix for Issue #3257 and 3285
The JBI runtime configuration page does not handle multiple targets when running
under the cluster profile.  Their needs to be a drop down added to the page that
allows the user to select the target they would like to configure.  There is
also a problem with the drop downs for the components logger, configuration and
monitoring page.  These drop downs should include cluster instances, since JBI
allows the user to configure and monitor components on an instance bases.

Added support (i.e. drop down) for the runtime configuration page.
Also, updated the drop down for the component logger, configuration
and monitoring pages to include cluster instances.

Reviewed By: Mike Write

Tests-run: QL, jbi-test