Issue number:  3315 Obtained from: issue tracker Submitted by:  byron Nevins Reviewed by:   sreenivas munnangi
In start/stop/delete cluster I call getAutoHADB() as usual.  If it returns true, we do HADB lifecycling.

If it returns false, we do nothing with HADB

This bug exists because an old assumption is no longer true -- the assumption was that all cluster-capable glassfish installations had HADB installed - or at least *should* have HADB installed.  Thus a severe log message is appropriate when HADB is not installed.

This is no longer the case.  The clusters profile uses in-memory replication.  The issuing of a harmless severe log message about HADB is bad.

MY FIX:
Only ClustersConfigMBean changed.  I still call getAutoHADB() but the implementation of getAutoHADB() now checks for the persistence type in the cluster config.  If the type is not "ha", then getAutoHADB immediately returns with false.  It doesn't call the HADB MBean at all -- thus no HADB log messages of any kind are emitted.