Issue number:  6581651 Obtained from: bugtraq Submitted by:  byron nevins Reviewed by:   anissa lam
Every server instance that is started has its stdout and stderr "drained" to avoid hanging.  Sometimes this output
is redirected to the server's server.log file.  When these threads died they did not close the log file's stream.

Fix: add a finally clause that closes the logfile's stream which unlocks the file on Windows.
Ran manual tets and QL-PE, QL-EE