Fix for Issue #3385
Changed the JBI logger page to have a "DEFAULT" option for each logger.  This
option will pass the null value for the log level to the JBI common client.  The
common client method will then reset the logger to inherent its values from the
parent logger.

Also with this commit, I changed the Log Level values being displayed in the
pull down to be I18n'ed.

Reviewed By: Mike Write

Tests-run: QL, jbi-test