Issue number:  6519530 Obtained from: Submitted by:  Sivakumar Thyagarajan Reviewed by:   Binod PG
Partial fix for 6519530.
We now create a new classloader chain "addon chain" [whose parent is set to the shared chain] to load all addons and provide a public API to uninstall an addon. When uninstallAddon is called, we remove the addon's classloader from the chain, so that the relevant jars could be deleted by the addon's uninstaller.

Tests run: QL, CTS smoke, connector devtests, SQE JMS