Issue number: 3416 Obtained from: - Submitted by:  Sivakumar Thyagarajan Reviewed by:   Byron Nevins
In ASLauncher, we append empty strings for libClasspath and classpathSystem in the system classpath to yield something like this:
java -cp "appserv-launch.jar:::".

The empty Strings in the classpath are considered as "current directory" and added to the search path for getResourceAsStream() in the system classloader and hence an app could load domain.xml as mentioned in 3416. Ensuring now that empty strings do not get added to the system classpath of the appserver process.


Thanks to Harsha and kumar for helping out in the investigation.

Tests run: QL, CTS smoke, SQE JMS, Connector and classloader devtests, manually tested loading domain.xml via localhost:4848