Fix for Issue 3192 - SUNWappserver 9.1 beta HTTP 405 Error on webservices (b41d b
Problem: Web services do not work correctly from a remote instance

During deployment, the app's descriptor is serialized by the DAS to a file in the app's "generated" directory so that loading during later starts can avoid reparsing the XML. This same serialized data file is copied to any remote instances to which the app is deployed during synchronization.

The serialized descriptor contains DAS-specific paths in the URLs for the WSDL files in the app.  This is because the ModuleContentLinker populates this information during deployment.  In the remote instance during loading by the WebModulesManager or EjbModulesManager, the serialized descriptor is read and again the ModuleContentLinker is used to fill in any missing URLs in the descriptor.  This time, though, because the URLs are already set (they were assigned by the DAS, serialized on the DAS, copied by admin to the remote instance, and deserialized that way) ModuleContentLinker does not change them, so they still refer to the DAS path.

Fix: Provide a way so ModuleContentLinker can optionally force the WSDL URLs to be updated even if they are already assigned values.  When the WebModulesManager and EjbModulesManager instantiate the ModuleContentLinker objects they use, they specify to force URL assignment.

Reviewed by: Hong, Bhakti


Tests:

PE SQE quick-looks; deployment, ejb, and web services devtests pending (thanks to Bhakti for the ws testing)


Code changes:

appserv-commons/src/java/com/sun/enterprise/deployment/util/ModuleContentLinker.java
appserv-core/src/java/com/sun/enterprise/instance/EjbModulesManager.java
appserv-core/src/java/com/sun/enterprise/instance/WebModulesManager.java