Bug Id: 6586598 Submitted by: Harpreet Singh Tests Run: PE/EE QL, devtests/callflow Reviewed by: Self
Synposis: Derby engines allows to create a table with - in its name but the PreparedStmt XML par
sing engine barfs on this character.

Replace a table name with - or . in the name with an _.
The table name is created of the server instance name. This issue was creating a problem when a
server with name ABC-1 was created and callflow was turned on for that instance.

Files Modified
src/java/com/sun/enterprise/admin/monitor/callflow/AbstractTableAccessObject.java