Issue number:  3453 Obtained from: TopLink Submitted by:  Tom Ware Reviewed by:   Guy Pelletier
Avoid creating a temp loader when static weaving is enabled.

Tested with entity-persistence-tests and QuickLook

An automated test was not added because this issue only manifests itself when running on the GlassFish container and we do not currently have a testing framework based in the container.

Bug 3452 was added as a request for this kind of framework