Issue number:   6436680 Obtained from:  Sun's Bug Database Submitted by:   L10N team Reviewed by:    Self
Made sure that a localized index.html gets copied during asadmin create-domain.
This *much* improves the usability of the product in case:
- the domain is created in non-US locale.
- index.html is available in the given locale. This can be easily achieved by
  copying the index.html in
  install-dir/lib/install/templates/locales/<locale>
  An example of locale is "fr_FR" or "zh_TW". The directory structure is
  communicated to the L10N team.

The actual dynamic dispatch of static resources based on *client's locale* is
still an RFE.

Tested *thoroughly* on various locales. The default en_US locale is not affected.
Will be available in 9.1pe-b58.