Bug: 6560079 Submitted by: Byron Nevins Reviewed by: Kedar Mhaswade
<admin-ee/admin> com.sun.enterprise.ee.admin.clientreg.JMXConnectorRegistry

changed method: isMBeanServerUnreachable()
added method: isUnknownHostException()

Summary:

isMBeanServerUnreachable unrolls the cause exceptions inside an Exception that is caught trying to connect to an MBean Server.  It looks for
ConnectException and InstanceException -- if either of those are found it returns true.

The problem is that if the server is down or has a bad host name or DNS isn't working right a java.rmi.UnknownHostException is thrown.  I added
code (isUnknownHostException()) to check for it.  Now if that Exception is in the list isMBeanServerUnreachable() returns true.

Tests: Manual tests (set NA client-hostname to garbage, set NA JMX port to garbage).  PE/EE QL