Fix for 6590921 ("Web Core must use default port for welcome-page redirection when PORT not part of http host header"):
If the request URL does not end with "/", and the container issues a
redirect, with "/" appended to the original request URL, we need to check
whether the original request has been proxied by a LB, and whether it
has been secure.
This info is important in order for the container to be able to
construct the correct redirect URL, which must point to the LB host
and the LB's correct (i.e., secure vs. non-secure) port.

Reviewed by: jfarcand
Approved by: mark.basler@sun.com

Tests run: QL