Trivial change.  There were 2 methods that did this:
rmiClient = getRMIClient()
.... use rmiClient once ...

-----------
rmiClient is the name of the instance variable.  Thus

rmiClient = getRMIClient()

-- the local rmiClient is set to the instance rmiClient and hides it.

This is a harmless condition and not a bug but it is dumb and generates warnings.
Since rmiClient is only used once in each of the 2 methods, I got rid of the local variable and use getRMIClient() directly.

Thanks to Kedar and NB6 for finding this!