Issue number:   6595919 (Sun's internal database) Obtained from:  Bug database Submitted by:   Mark Basler Reviewed by:    self.
Patch provided by Peter Williams. This is a hack and should be handled
with care. This is done only to make progress on AIX builds so that we
can create clusters on it. The issue is with the generated code relying
on attribute "name" rather than "Name".

Interestingly, it is revealed only on another implementation -- IBM's JDK.
Fix is an acceptable hack because it is localized. Maintainers of this
code should note that every time this file is regenerated, the patch should
be applied again. Most likely, DTD won't change for 9.x GlassFish releases,
so we should be OK.

Tested with quick-look by creating a remote instance. Tested creation/startup
of a cluster as well. Tested on Mac OS X.