Fixed broken build.
Change the ResourceAdapter binder from an Init service to a lazy apprach when the first initialContext 
is created.
fixed some unecessary injections
change the Sniffer lookup to actually happen only when an application is deployed or started.
changed the mbeanserver initialization to Async

Added back the gf-web-connector that contains only the sniffer.