PlainTextActionReporter.java --> A new ActionReport that writes plain text to the output stream to the client.  This would be surprisingly difficult to do using an existing ActionReport implementation.
XMLActionReporter.java --> It declared @Override on a method that used to be abstract in ActionReporter but is now part of the interface.  Thus it
is illegal to have @Override

The other file changes:

Before the AdminAdapter created an ActionReport that was based on the user-agent of the request.  There was no way for an admin command downstream
to change the reporter.  I added man page support and need a PlainTextActionReporter *if* the man page is available.  If it is not available the default
ActionReporter subclass is used for generated man page info.
Now the ActionReport object is passed downstream and a ActionReport is passed back upstream.  So now called code can create a new 
ActionReport and have it take effect.

Man Pages are added into the jar and are returned with "--help" commands.

Also AdminAdapter now works with a ActionReport instance -- not an ActionReporter instance.  This allows one to implement the ActionReport 
interface WITHOUT being forced to extend ActionReporter.