IT4624
the streaming XML parser's close() method does NOT close the underlying stream.
As a result, domain.xml was left open.

I moved the domain.xml InputStream from a method field variable to an instance variable. 
Both the parser and the strem are explicitly closed in a new finally block in the constructor, so closing is now guaranteed.

Reviwed by Sreeni