1. Added a new module called jxmremote_optional under distributions/external. This is responsible for repackaging of jmxremote_optional.jar. Changed AMX and admin/jmx* stuff to use the repackaged artifact. After this change, we no longer have to add jmxremote_optional.jar to classpath in ASMainOSGi.
2. Added two new modules called commons-codec and sysnet-registration
under distributions/external. They are responsible for repackaging 
commons-codec and sysnet-registration jar as modules. Changed registration-impl
and admingui to make use of the new repackaged jar. 

3. Fixed a missing module dependency from admingui/plugin-service to hk2 module.
More over, the dependency on glassfish-api was marked as provided - that's not
considered by our hk2-maven-plugin to generate OSGi Require-Bundle header. 
Removed the scope.

4. Since we repackage classes inside apache-commons module, we don't have to
distribute the original jars. Hence marked those dependencies as optional. For
code that use apache-commons to compile fine, we are unpacking the classes in
target/classes as well.

5. extras/phobos used to depend on commons-logging, as a result it was causing
commons-logging.jar to be part of distribution. Since commons-loggin has been
wrapped as part of apache-commons, I am now setting the dependency on 
apache-commons.

Reviwed by: Llyod, Kedar, Siraj, Jerome