[fixed] issue #4713
Changed the distribution assembly logic so as not to handle plain jars and hk2-jars differently.
The original reason for doing that was due to the way HK2 arranges jars and Class-Path manifest
entry, but in the OSGi world where every module is fully self-contained, this is no longer
necessary.

This was causing duplication of multiple jars (javax.javaee-*.jar was one of them), as those
jars were generated with <packaging>jar</packaging> instead of <packaging>hk2-jar</packaging>.

This also simplified the magic build.xml script.