I spent 2 hours poring over this extremely complicated code to find out what the difference is between a deploy command that uploads a file and every other command.  The code showed what appeared to be very very complex differences.
In reality -- it is a TRIVIAL difference.  If you access & write to the stream of the http connection -- JDK makes the request a POST.
If you don't write to it -- the request is a GET.

So -- the ONLY difference between a deploy with a file to be uploaded and all the other commands -- is you write the file to the connection.  Period.  
EVERYTHING else is exactly the same.

I changed the code around so that other developers (or me) don't have to waste time figuring this out again.

There is still lots more work left to get this way over-complicated code tamed...

FileUploadUtil -- not needed.