things like DigesterFactory should be a component.
The point of making something a service is so that it can initialize itself before it gets used. The way it was done before, code like this always had to get initialized by someone else, and therefore it breaks the basic encapsulation principle.

I wanted to do the same for ParserUtils, but it's been used in too many places so I abandoned it for now.