Merging the embeddability enhancement in branches/kohsuke/post-tp2, from rev.19655 to rev.20792. Aside from various minor small cosmetic refacotring changes everywhere, the main change is around making several features full HK2 services, so that they can be swapped in the embedded environment.
In several places, I had to rely on the "Globals" class hack to access habitats, since the caller
of those pluggability points are not HK2 components by themselves and therefore couldn't request
an injection, which is the proper way of doing this.

"Globals" in this way creates a front line of properly component-ized world and the classical world.
We should gradually try to push this boundary further, so that more of the system is modular.


This merged workspace passed all QL tests in both HK2 mode and Felix mode.