No longer depend on web container to call GrizzlyService.registerEndpoint().
This should no longer be required, now that ContainerMapper uses the
listener's associated Mapper, which has been populated with any web
endpoints from corresponding jmx mbeans (which are registered whenever a
host, context, or servlet is started).

The only reason calls to GrizzlyService.registerEndpoint() from the
web container were still required was so that ContainerMapper could
capture the (Coyote)Adapter passed to GrizzlyService.registerEndpoint().
Added getAdapter() and setAdapter() methods to Mapper, so that the
(Coyote)Adapter may be retrieved from it.

Tests run: QL