Fix 4974 - dependence on hard-coded file types in autodeployer
In v2 GF knew what file types to look for in the autodeploy directory.  In v3 the autodeployer does not know ahead of time how to identify legitimate apps.  So these changes basically remove the file type checking and simply try to deploy each file that's in the autodeploy directory. So many files are affected because of some other clean-up, esp. of directory autodeployment and undeployment.

We may want to come back later and improve some of the error reporting but it's passible for now.

Reviewed: Hong
Tests: QL, manual autodeploy tests of valid and invalid files