Add logger changes per Jeromes email (see extracted content below).  This change also fixes bug 5455. API change was applied to all that called LogDomains.getLogger.  LogString.properties file was moved to the appropriate locati on in order for the quicklook tests to pass.  More need to be moved.  See email subject: URGENT and ACTION:   changes to logger code will affect you.
From Jerome's email
...
I am proposing we change the API to
     Logger getLogger(Class requesterClass, String loggerName);

The role of the Class parameter is to obtainer a class loader doing requesterClass.getClassLoader() which will be responsible 
for finding the resource bundle. The name will just be used to identify which logger instance you request.

the resource bundle will be looked using the following algorithm

   1. we will get the Package name from the passed requesterClass and lookup the resource bundle in that package using the fol
lowing name :
            requesterClass.getPackage().getName + ".LocalString.properties" and the current Locale of course.
   2. we will walk up the package hierarchy and repeat (2) until we get to the root
   3. for backward compatibility with existing code, we will use the obtained class loader to load the old resource bundle usi
ng the mapping we had in v2 (this lookup will gradually fail more and more as people move their resource bundle in their packa
ge).
   4 throw a MissingResourceBundleException.

It does mean that if you start using  a LocalStrings file in your module's packages, we will not load the resources from the p
revious resource bundle as in v2 and you need to move these resources in your module.

Final note:  Changes affect LogStrings.properties and not LocalStrings.properties.