Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=6163 ("REGRESSION] http://host:port/ doesnot return /index.html page")
The following scenario caused a ClassCastException inside the WebContainer:

http://localhost:8080/
http://localhost:8080/index.jsp //start the WebContainer on the fly
http://localhost:8080

Since Grizzly's Glue Code first register it's StaticResourcesAdapter (for *.html), the second request cause Grizzly's Glue Code to replace the RRO mapoing with the SnifferAdapter, which in turn get replaced by the CoyoteAdapter. When the CoyoteAdapter take controlof the ROOT, it doesn't remove the previous mapping (nor replace it) but instead happens its context/wrapper (Servlet) to the list of mapping. Hence when the next *.html request arrive it gets mapped to the CoyoteAdapter and the first Wrapper, which is not the DefaultServlet but instead the StaticResourceAdapter. This is what causing the ClasscastException. The fix consist of unregistering the ROOT so previous mapping are all gone.

- ALL QL passed
- Scenario above now pass.