GlassFish: 6351
Made domain creation possible even if the keytool
could not be found. If keytool is not found or there is
some error in running it, keystore.jks and/or cacerts.jks
are not created in domain's config folder.

This is ok for now for the reasons described in the bug report.

With no keystore/trust-store, QL fails miserably with 23 failures.
With JDK at the time of create-domain and properly created key/trust-stores,
everything is all right.

Note -- we are NOT trying to say we run with JRE. There are other problems.
It's only to make sure we don't look ugly on installer -- see 6286.