IT 6126
Even though the cause from serverside is ignored if AS_DEBUG=false, we still need to save it some where.  So save the cause from the serverside in RemoveException.  The cause from the server side is used to check if command is invalid.  If command is invalid then call displayClosestMatch.

passed manual and QL tests.