Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=6158 ("Class Cast Exception during startup of domain")
The current issue occurs because when http://localhost:8080 is called, grizzly-kernel is creating a MappingData associated with / . All requests to context path / are mapped to Grizzly Static Adapter. When the WebContainer get started, it replace the / but never remove the wrapper associated with the previously added MappingData. Next request cause the ClassCastException as the request is properly mapped to the WebContainer, but the MappingData contains invalid information. 

+ All QL passed
+ Manually tested the scenario described in the issue

Reviewed by Jab Luehe and Oleksiy Stashok
Approved by Jerome Dochez