Changed vis pkg-private --> protected for Server.decorateInhabitantsParser() as requested:
Hello Byron,

I have what I hope is a small embedded Glassfish change request: can you restore the 
decorateInhabitantsParser() method in embedded.Server to protected instead of package private?

I know this is out of the blue, so a little background:  I work in Sun Labs on Project Wonderland 
(http://wonderland.dev.java.net).  Our next version uses embedded Glassfish as the basis for 
a web-based management console that we ship with Wonderland.  Embedded Glassfish has works 
great for us -- rather than making the user manage all the processes and configuration data 
for the virtual world, they can now do it with a web UI.

Currently, we are based on an old version of embedded Glassfish. In this version, we found 
we wanted to customize a number of different things.  Extending Server to do some additional 
decoration of the inhabitants parser was definitely our most powerful tool.  As a concrete example, 
we need Wonderland to work without Internet access, so we used decoration to replace the 
WebArchivist and WebDeployer with versions that have the "load-external-DTDs" flag set to false.  
If it is possible to make that method available again, that would be quite helpful.

Thanks!
      -Jonathan

--------------------------------------------------------------------------
Jonathan Kaplan
Sun Microsystems Laboratories
Menlo Park, CA
http://research.sun.com/projects/dashboard.php?id=85
jonathan.kaplan@sun.com