IT 56
AutoDeploy Service is removed if and only if 
EmbeddedInfo.autoDeploy == false
false is the default so user needs to do these 2 steps to enable AutoDeploy

(1)  info.enableAutoDeploy()
(2) Supply a domain.xml that has the following:
     <admin-service>
          <das-config autodeploy-enabled="true" ...... />

Embedded does nt use MiniXmlParser yet so I did not want to add the overhead of it for this
unlikely-to-be-used feature.
If we ever *do* add user-supplied domain.xml parsing then we should make sure that the following
conditions are true before enabling the service:

(1) the autodeploy directory exists
(2) autodeploy is enabled.