IT#71
The java.ext.dirs setting was totally useless.  I think perhaps dirs that are so
listed need to be in a policy file or something.

Here is what I did:

1) got rid of the java.ext.dirs stuff
2) Add 2 jar files to the classpath -- derby.jar and derbyclient.jar

if either of these files can't be located the launch will fail with a helpful
error message.

I also added a test to Server.java.

It logs a message reporting whether or not Derby is available (i.e. it tries to
load the driver class)