We found leaking logger objects. For a large perf run there were more than 2 million logger objects referenced in the heap dump.
The cause was:

LogHelper.getDefaultLogger() is called for every requests and this in turn calls Logger.getAnonymousLogger(). Each call to Logger.getAnonymousLogger() results in to a new WeakReference added to Logger.kids array. So for a million requests you would have a million Logger objects in Logger.kids and these do not get GCed as the root logger remains till JVM exits.

To fix it, VsAdapter has it's own logger as private final. LogHelper.getDefaultLogger() is refenreced from few other places in v3 code. It should return a singleton logger instead of creating a new anonymous logger on each getter.