Enable java2db support in JPA and CMP.
- Added clean(DeploymentContext) method to CMPDeployer
- All intermediate values are stored in the DeploymentContext properties, which are available in clean, and
do not require to reload DOL objects or persistence.xml.
- Set params.origin for EJBTimerService auto-deploy to that of the user app that causes it to be deployed.
This allows to treat auto-deploy as load for the java2db purpose on server restart. The only exception
is the server restart after all user applications with timeouts had been undeployed. This doesn't cause any
failures, but only a warning in the server.log.

Reviewed by Mitesh. QL and manual tests passed.