Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=6935 ("No optional client authentication available in HTTPS connector")
This time test in GlassFish and Grizzly, not only Grizzly. Thanks for bbfish for its great analysis of why it failed in v3 but not in Grizzly. Now just jar your SSLImplementation and just drop it under glassfish.home/lib and v3 will be able to load it. 

The best solution would consist of creating an OSGi module with your class/module.