Incremental Fix for 7656 - Misleading deployment errors if directory deployed app's submodules are not expanded
The additional fix added here handles directory deployments, in which submodules do not appear in JAR form but in expanded directory form.  The app client deployer was not correctly handling this case, with the result that the downloaded app client files were incomplete and did not work.

Tests: QL, ejb and deployment devtests