Fix for issues 8390 and 7438 Porting to J1 Branch.
8390: This fix is to skip OpenInstaller from trying to initialize certain UI elements even in
silent installation mode. This is visible as JDKPanel is suppressed in Mac due to an open issue
with OI in detecting JDKs on Mac.

7348: Workaround for updatecenter notifier bug on Mac. On Mac, the command 'updatetool --notifier --shutdown'
does not work. The explicit call to this command during uninstallation is removed as a result of this
fix and till the issue is fixed in updatecenter, users will be instructed  in install guide to stop
related processes prior to uninstall.