You know those CLI properties arguments that look like this:
foo=goo:abc:hello=:x=y

This class will convert them into a Map<String,String>
It is painful and difficult to parse this stuff out and think of (and handle) every possible way that the user can foul it up -- thus this centralized class.  We don't want to write this code more than once so common-utils is its home.