GLASSFISH-8471
The issue that keeps on coming back -- installing in a location with spaces
in it breaks asadmin. Hopefully, I have now taken care of it ...

Tested on paths with and without spaces, on Linux only. Waiting to install
VB image to test it on Windoze, where I think the bug is taken care of, already.