Removed some logging code that was intercepting log messages and outputting them manually rather than using the logger. LogService now has a no-arg getLogger method so that all the classes don't have to pass in the same string to all get the same logger. LogService was also performing some unnecessary tasks every time a logger was retrieved.
Cleaned up some imports along the way.

Made some loggers and string manager fields static and final, and removed some public
constructors that were therefore no longer needed.

With this change, the log information is now going to an a file "upgrade.log" in the current
working directory rather than going to standard out. This can be easily changed by a
template field in LogService.

Note: with the logging now going to a standard place (instead of having different
'listeners' in the code sending it where they want), it is not showing up in the
GUI progress screen. Am committing anyway so that I can work on the GUI implementation
with a clean slate.