Adding support for a new feature of JDBC connection pool : java.sql.Driver based pooling support.
This is a proposal to support configuration of JDBC drivers and applications that use java.sql.Driver implementations.

The res-type attribute when set to java.sql.Driver enables the Driver manager based MCF. A new attribute driver-classname will take implementing driver class name as the value. driver-classname will be a mandatory attribute when the res-type is set to java.sql.Driver.

datasource-classname attribute will be an optional attribute when res-type is set to java.sql.Driver, whereas the driver-classname attribute will be mandatory. 

Reviewed by : Jagadish Ramu 

Tests run : QL, Connector/jdbc devtests, tests related to DMMCF.