Adding support for a new feature of jdbc-connection-pool and connector-connection-pool : Ping Pool during creation.
The connection pool will be automatically tested for valid connectivity, when it is created by setting a new attribute ping to true. This feature is useful to identify erroneous values, if any, specified in the jdbc connection pool configuration.

The existing pooling infrastructure does not warn the user of any unsupported/incorrect value of a database connectivity property while it is created. Only at runtime, the pool is initialized with the properties and attributes specified by the user and while getting a connection a warning would be seen.

Default value of ping is false which indicates that the pool will not be tested during creation. The existing "Ping" button in admin GUI and the "asadmin ping-connection-pool" command would exist and be supported in the future.

Reviewed by : Jagadish ramu

tests run : QL, connector/jdbc devtests