Fix for issues 8995, 8996, 8987.
Introducing a semaphore in the ListDataStructure to be used while creating threads. This is needed in case of concurrent threads accessing the pool. Also, when notification happens in ConnectionPool to enable removal of an element from the wait queue, the waitMonitor is got by removing an element from the waitQueue instead of getting the first element. 

This fixes the incorrect connection request wait times in case of multiple concurrent threads accessing the pool. The connection creation count is also updated with the right values as per the max pool size. 

Tests run: QL, JDBC/Connector devtests, appserver-sqe tests for jdbc monitoring.