GlassFish: 7724
Fixed generate-jvm-report to show usage text when presented with a wrong value for type.
In the process made changes to report such that for a given instance of ActionReporter,
the overall report will contain both message and failure-cause only if they differ. This
takes care of duplicate display in case of errors, because server sometimes does:
report.setMessage(exception.getMessage()) and
report.setFailureCause(exception) in case of failures which results in duplication.

QL passed.