Issue: 9313 This will be an important feature. Consider a piece of code in glassfish workspace, where we want to define a probe so I could get some important stats. The issue here is that, the provider method which we are designating as probe, would consist of parameters that might not be as useful for the probe listener. It would be important to massage these parameters before I would fire the probe with more meaningful data for monitoring.
We do this in a two step fashion. First we define a probe in XML for a method in
code where we wont necessarily have access (consider Eclipse link here). I would
like to mark this probe as hidden. Now I write a listener for this probe in the
module where I have access to code (consider JPA) and then massage the
parameters I recieve. Once I have the meaningful data at hand, I will fire a
probe at this point, for this I will designate a new method as a Probe (this is
the visible probe for the end user).

Also note that, to make some sense of the first probe data, I might need access
to the 'this' object of the provider (Eclipse link's probe provider object). We
do this by denoting 'self=true' in the probe definition in XML.

This I think will be a very useful feature of monitoring.