Made the LDAP authentication (not authorization) work for admin CLI. This is an important step toward being able to use corporate LDAPs for administration. Admin console being a web-app is already able to support that!
What's still not working is groupm memberships from LDAP. Thus, at the moment, any
random LDAP user will be able to administer any domain. This should soon be fixed.

Modified create-file-user to warn users of anonymous access when a new admin
user is created.

QL passes 100%. Lots of local testing, especially using LDAP. Things are good.

Thanks to Sahoo who suggested that in order to use the security life cycle,
server's thread context class loader should be CommonClassLoader, which it is not.
That might be a separate bug. Because of that, set the ContextClassLoader to
CommonClassLoader (from ClassLoaderHierarchy) if it is not set. Used only during
LDAP authentication.