Partial fix for 9413
The server was logging a warning from the Java Web Start support because a cmp devtest's EAR did not contain a JAR referenced from the app client JAR's manifest Class-Path.  (Such warnings are desireable because the absence of a JAR specified in the client can cause different behavior in the client from what the developer or administrator expects.)  Although the warning was correctly issued, there was a separate problem in the extension JAR checking logic which this check-in fixes.

Even so, the cmp dev test will need to be changed to avoid the warning.

Tests: QL, deployment and EJB devtests (no additional or different failures compared to the EJB devtest failures using the nightly build)