Fix for 9610, performance degradation of cli due to hk2
The degradation was due to cli-optional.jar which has a class-path setting to import the entire
set of jars we deliver. Removed that classpath setting reduces the asadmin cli start to 1/3. 

Reintroduce support for DomainInitializer which was turned off due to build failures. Rewrote
it to use the inplanted mode (hence not relying on huge classpath).