Fix for issue 8977.
The fix introduces a reset() method annotated with @Reset in the stats providers. This is called when monitoring is set to HIGH from OFF. Also, during pool initialization, the numconnfree count is set to the steadyPoolSize. This ensures that the pool counter numconnfree is initialized with the right value before any updates to the statistics. The reset method updates numconnfree and numconused based on the pool's current state.

Reviewed by : Jagadish Ramu
Tests run : QL, jdbc devtests