Issue 9559 : some configuation changes are not applied.
this was actually an invalid bug, the originator was trying to modify a plain list object returned by 
a @DuckTyped method thinking he was changing the live configuration list.

I made all @DuckTyped methods that return a list or a Map to return an unmodifiable version so 
users get an exeception when adding or removing elements.