GlassFish: 9340
Don't let the last definition of a system-property to be deleted
if there are references in the configuration. Thus a system
property PORT will not be allowed to be deleted if some attribute
somewhere says ${PORT} as its value. 

Beware: The process results in recursive search for all attributes of
all elements in domain.xml. This was discussed with Kohsuke.