IT 10255
Once upon a time start-domain insisted that if a passwordfile was given as an arg then it MUST contain username, password and master password.
That is no longer true.  Now a password file can contain 0, 1,  2 or 3 of those properties.
Changed create-service code accordngly.

We still hand the gun to the user so tht he can shoot himself in the foot, but we do NOT pull the trigger for him.
I.e. he can have bad passwords or format in the passwordfile -- heck he can delete the file later or whatever...