https://glassfish.dev.java.net/issues/show_bug.cgi?id=9934
JDK6 bug.  Discovered by trying to build with JDK7

Superclass method is trying to access a private variable of superclass through a reference to a subclass.  This was never allowed in Java.
The compiler in 6 must be confused by the Generics

Fix:  Simply upcast the subclass reference. (Suggested by Bill Shannon)