At least a partial fix for 10630
The implementation of the entries() method on the InputJarArchive class collected all entry names into a Vector before returning an enumeration of that Vector as the method value.  For a large JAR this can require significant time and memory.   This change converts that logic to return an enumeration that does not automatically load all the entry names into memory.   

Tests: deployment and ejb devtests, QL