Fix for 10732
The drive letter on Windows needs to be upper-case for security to work (!).  Depending on how the appclient script was invoked, the drive letter might be upper- or lower-case.  It must be upper-case.  This change forces the drive letter to upper case (WIndows only).


Tests: deployment and app client devtests; Sherry's help with the failing sqe tests