Fix for 10807
The ACC's management of the class path using a masking class loader caused some security tests to fail.  Because this particular change was not essential to the major appclient use cases I am reverting it, at least for now.  Later on we can revisit this.

Tests: QL, deployment and selected EJB dev tests; the sample apps attached to 10535