Fix for CR 6903580
The app client deployer logic which records what files should be downloaded (as part of "deploy --retrieve" or "get-client-stubs" or a Java Web Start launch) incorrectly requested for directory files that are referenced from JAR Class-Path entries to be downloaded.  It should have requested only that the non-directory contents of such directories be downloaded.  The deploy command or get-client-stubs commands' attempt to write the directory file into the payload of the http response would fail, aborting the download.

This fix continues to record that such directories have been processed (to avoid processing them multiple times if multiple Class-Paths refer to the same directory, for example) but it does not include the directory files themselves in the downloadable artifacts...only the contents.

Tests: QL, deployment and ejb and selected webservices devtests; Java Web Start tests; selected cts tests which had exposed this error
Check-in approved by Jerome