Fix for 11126
Deployment of an EAR which contains a submodule below the top-level (for example, xxx/myEJB.jar) fails on Windows.

There was a subtle difference in how the FileArchive logic handled / vs. \ in the module's URI, compounded by a subtle difference in how the EarDeployer invokes createSubArchive (when it expands the submodul into the expanded directory) vs how it invokes getSubArchive (which it later invokes to get the expanded directory name).  

See the issue for the full, sordid details.

This fix causes makeLegalFilename in FileUtils treat / and \ equally when it processes them in submodule paths.

Tests: QL, deployment and ejb; appclient test that had exposed the problem
Check-in approved by Jerome