Fix for 11181
A problem occurs when all of the following is true:

- the app is an EAR which contains an app client
- the EAR is directory-deployed
- the EAR (say it's name is myApp) contains an app client module myAppClient.jar (which, because this is a directory deployment, is pre-expanded by the developer into myAppClient_jar)

Attempts to launch the app client fails with

    java.lang.ClassNotFoundException: (main-class-for-the-client)

This is because the AppClientGroupFacadeGenerator placed the generated group facade into the app client's generated/xml directory whereas it should go into the EAR's generated/xml directory.  As a result, the group facade would overwrite the JAR that is generated from the developer's expanded app client JAR directory.  The first manifestation of this is the inability to find the main class from the client.  

With this fix the group facade is placed into the EAR's generated/xml directory.  Also, a related change is that the DIRECTORY_DEPLOYED property is set in the deployment context's deployment properties for redeployment so that the app client deployer can use a single place - the app properties - to tell if the app was directory deployed or not.

Tests:  QL, deployment and ejb devtests; Elena's directory-deployed timersession app attached to the issue