Fix for 11218
Directory deployed apps which contain an app client would not load correctly during a server restart, because the deployment context's app properties did not include the directory-deployed property setting if the app is in fact directory deployed.  In this change, the DIRECTORY_DEPLOYED property is set in the deployment context's deployment properties for redeployment (as well as deployment, as before) so that the app client deployer can use a single place - the app properties - to tell if the app was directory deployed or not.

Tests:  QL, deployment and ejb devtests; app attached to the issue