Handle the rare case in which the LogManager reports a logger name but later returns null when that logger is requested.
(For historical reasons, the ACC modifies all known loggers if necessary so that each logger level is at least as detailed as the configured logging settingin sun-acc.xml.  It does so by asking the LogManager for all logger names and then asking for each logger by name.  A user reported a case in which such a returned logger was null.)