Fixes issue 11287
When it launches an app client the ACC performs various checks on the client archive, such as making sure the client has a manifest and that the manifest specifies a Main-Class.  The error handling logic in such cases did not correctly format the name of the offending archive which led to a URI-related error.

These changes correct the handling of the URI so the file name is correctly formatted without a URI error.