Fix for 11339
Java Web Start launches of directory deployed app clients (nested in EARs or stand-alone) would fail because Java Web Start discovered an unsigned entry in a JAR file that should be signed.

The problem was that the auto-signed copy of the generated group facade JAR was not written to the correct location.  When Java Web Start would request the JAR GlassFish would serve the unsigned copy.

The basic fix is to make sure that GlassFish generates the auto-signed generated group facade JAR into the correct location.  Some of the other changes implement some improved error detection and clean up some of the related logic.

Tests: deployment and EJB devtests; QL; manual test of the example for 11181 and a NetBeans-deployed example app