Fix for 11644.
The API and the doc for the embedded ACC correctly says that the calling program should create the new container using 

AppClientContainer acc = AppClientContainerBuilder.newContainer(...)  

(after configuring the builder as desired) and then invoke 

acc.startClient(clientArgs)

Following the doc and the API generated an IllegalStateException because the startClient method implementation did not invoke the AppClientContainer.prepare method as it should have.  

This fix corrects that error.

Tests: deployment and ejb  devtests; QL