Fixes error reporting problem in the admin CLI client.
Earlier changes to the PayloadFilesManager introduced an error in how server-side errors were reported to the admin CLI client.  

These changes fix those errors.  (An error during report handling used to be wrapped in an IOException but that hampered the CLI client's ability to handle the report display.)

Tests: deployment devtests