Fix for 11861
This changes replaces each backslash In the file-xfer-root value that is sent with each transfered file to a forward slash.  This is because the file-xfer-root needs to be a URI (typically relative) so that the receiving end of the transfer can combine it with the name of the file Part in the response (also a URI) to compute the correct output location for the extracted file.

Tests: deployment devtests on both Windows and Mac OS X.  Also manual tests on Windows for a path containing a backslash.