https://glassfish.dev.java.net/issues/show_bug.cgi?id=11909
We get a half-dozen scary-looking SEVERE log messages that Felix spews out if the Felix Telnet port is already in use.  GlassFish itself does not check or comment on this.

I added a check in the Launcher to detect this condition -- and it is logged as a SEVERE message if applicable.