Fix for issue 11653
An error while we merge EntityManagerReferenceDescriptors in WebFragmentDescriptors led to missing of EntityManagerReference related InjectionTargets in the WebBundleDescriptor. 
This prevented Java EE style injection of EntityManager in the producer field of the bean

Thanks to Ken and Hong for helping in the investigation.
Reviewed by: Shing Wai Chan
Tests run: QL, manually ensured that the application in the bug report worked.