Initial implementation of remote deployment (deployment of an app to one or more remote instances).
The remote instance deployment logic is a no-op placeholder for the moment.  It simply logs the uploaded archive, deployment plan (if provided), and generated directories.  

The DAS side should be fully functional, triggering an "after" supplemental command which creates and sends a "hidden" remote deploy command to the correct targets, passing the original deployment command parameters plus additional ones to send the generated directories for the just-deployed app.

Tests:  deployment devtests, QL, manual test of deployment with --target=a-cluster