Bill Shannon found a fat bug by inspection:
------
I was looking at ProcessManager and I think I see a bug...

It's writing the stdin to the process before starting the
threads that drain stdout/stderr.

If the process generates a lot of data to stdout before
reading stdin, filling the pipe, it will block waiting
for the pipe to drain.  Meanwhile, if ProcessManager
writes a lot of data to the process's stdin, filling the
pipe, *it* will block before it starts the thread that
drain the stdout/stderr.  Deadlock.

---

I moved the writing to stdin of the spawned proces to AFTER the reader threads are started.

It was easy to make it hang.  Simply write 500,000 chars to a process of "tee".  It immediately hung.