Part 2 of Issue 12604 Banish node-agent-ref from domain.xml
The PortManager needs to change to use the node host, instead of node name or node agent ref.  Added the getHost method from RemoteInstanceCommandHelper as a duck typed method on Server so it can easily be used by PortManager.  After ms3, we can remove getHost from as RemoteInstanceCommandHelper and change callers to use getHost on Server.