Fix for 4756 - a test requires an Internet connection
During JAXB class generation JAXB reads the DTD for the acc config file.  Although GlassFish (the server and the ACC) handle DTD lookups locally the test did not.

These changes allow the test to look up DTDs locally. 

The pom.xml changes include corrections to the scope of the dtds.zip dependency.  "Provided" is the right choice because the dtds.zip is used only during building the module and is not packaged with the module.

pom changes: approved by Jane
Tests: QL, deployment devtests