Fix for 12763
The generated and downloadable artifacts information was being added to an in-memory data structure during load, not during generateArtifacts and not persisted anywhere so it would be available after a server restart.  This caused get-client-stubs to fail to find the correct downloadable artifacts after a restart.

These changes accomplish two basic things:  First, the generated and downloadable artifacts are available from the DeploymentContext so Deployer.generateArtifacts and Deployer.clean can get the Artifacts objects from there and add artifacts to them or delete previously-generated ones as appropriate.  At the end of deployment, these two objects are extracted from the DC and persisted to application-level properties.  Second, the AppClientDeployer (the only deployer that actually needs to track downloadable artifacts so far) now uses this new approach to record downloadable artifacts.

Note that any Deployer that generates an artifact that it might need to clean up in Deployer.clean should track its generated artifacts by using the new utility methods in DeploymentUtil: generatedArtifacts and downloadableArtifacts.  

Tests: deployment devtests (DAS only and cluster), QL