Additional fix to allow off-line builds to work.
The earlier fix for this worked when build was run from the appclient/client/acc-config directory but not from any higher directory - such as the top-level, for example.

This fix uses a full path to point to where the extracted DTDs reside.  (The earlier fix used a relative path that worked during the build from that directory only.)

pom change reviewed by Jane