Fix problem sending a payload to a file.
Part of generate-sync-bundle's job is to create a file containing, well, a sync bundle.  The command uses the Payload framework to do so, not to send data in an HTTP request or response but just to store it into a zip file.  But the resulting files were unintelligible to the unzip utility.  This change fixes that.

Tests:  QL