Additional check-in for secure admin implementation
Refactored some logic from CreateServiceCommand that is now shared with CreateLocalInstanceCommand.  

CreateLocalInstanceCommand now passes more info to SecureAdminClientManager to construct the domain config if a domain.xml is available, given the parameters on the command line.  (If the user specifies a node or nodedir or server name the command now will try to build the domain config so it can see if secure client admin is enabled, etc.

Tests: QL