Fix for 13101.
Adding a file recursively to the payload caused PayloadImpl to open many file handles as each file was added, not as each was opened.

This fix uses the alternate form of attachFile which accepts a File (instead of an InputStream) and opens the stream only when needed to actually add the file's contents to the payload.

Tests: QL, Tom ran some separate cluster tests as well that had previously exhausted open file handles

Tests: QL, error scenario in the issue