start-instance: don't try to start an instance that is already running Fix null pointer exception in SSHLauncher found by logging team. Fix 12344: start-cluster/stop-cluster: make multi-threaded
For 12344 The number of instances operated on in parallel is
limited to be no more than 1/2 of the admin thread pool size.
Currently the default admin thread pool size is 5. This means,
by default, we will only operate on 2 instances in parallel.
For large clusters you will want to increase the admin thread
pool size. Tom reports that with a large thread pool he was
able to start a 24 instance cluster in 1 minute (33 secs
when no synchronization was needed).