Issues: 13087 - fileinstall properties are not set for an instance. 12285 - OSGi shell address in use message in instance server.log 12928 - need to put debug port in system-property in domain.xml 13113 - domain.xml contains v2-specific attribute
The OSGI_SHELL_TELNET_PORT and the JAVA_DEBUGGER_PORT are now part of the domain.xml template.  These are substituted at domain creation time for a domain and system properties are created for them in default-config for instances. The PortManager now assigned these ports. 

The fileinstall properties have been added to the default-config and the v2 GMS attribute has been removed from default-config for the domain.xml template. 

Ran QL.