Issue 13207 - add a separate thread pool for admin requests
This change introduces an "admin-thread-pool" to be used by the admin-listener. It has a maximum size of 50
by default with a queue size of 256. This is larger than the default of 5 for the http-thread-pool that the
admin-listener used before. 

The start-cluster code is changed to get the pool size of the pool used by the admin-listener, via the AdminEndpointDecider class.

Ran QL.