Fix for issue 13040
EJBContainerServicesImpl wasn't considering a super-class as a valid business interface for a no-interface
view of a session bean. Adding a method to EjbAbstractDescriptor to store/retrieve those supeclasses
for the bean and populating this list once in AbstractEjbHandler.

Tests run: PE and EE QL, EJB and CDI developer tests, and added a dev-test 
[cdi/javaee-component-resources/no-interface-ejb] to cover the scenarios listed
in the issue.

Thanks to Marina for reviewing the patch