fix for 13216 stop-instance/delete-instance mexed up about whether instance is stopped.
stop instance now tries to check if the server is down using the same algorithm as stop-local-instance which means it checks for the existence of a file in the config dir of the instance.  In the remote case if SSH is available I try to use it if not then we exit.