Create two new config bean validation constraints:
@NotTargetKeyword method/field constraint. Ensures a string is not a
reserved word. Right now the only resrved word is "domain". This
is used on the name getter for Server/Cluster/Config/Node and will
prevent you from creating an instance named "domain".

@NotDuplicateTargetName class-level constraint. Ensures a config bean
name does not clash with any other Server/Cluster/Config/Node. This
wll prevent you from creating a Server with the same name as a Cluster.
Those checks use to be done ad-hoc by each command itself.

Tweaked GenericCreateCommand to set the ActionReport message to be
the message from the root exception (instead of the top most exception),
but even with this error messages produced when these constraints are
violated are a bit ugly. I'll be creating a seperate bug to improve
messages for TransactionFailure exceptions.

Ran admin CLI devtests and quicklook.

Fixes issues:

13687 create-instance lets you create an instance named "domain"
12504 create-node-ssh must ensure name uniqueness across instances, etc.