Unrevert changes that were removed day before MS5.
Add gmsbindinterfaceaddress as an optional parameter to create-cluster.
For cases that one may want to set this value directly (such as to set to local loopback when all clustered instances run on one machine),
setting it on the commandline avoids an otherwise required restart of DAS after setting
this value after cluster is created with an asadmin set command.

Bobby reviewed change.