Further check-in for secure admin.
These changes allow asadmin and the DAS to follow HTTP redirection (now that we have the Grizzly config to do redirection so) if the server to which it sends admin requests redirects http to https.  Each redirection requires us to create a new URLConnection, so there is a little refactoring inside RemoteAdminCommand to allow the code which fetches the command metadata and the code which actually runs the command to prepare the connection as they wish before the main code sends the request to the new, redirected location.

Tests: QL, deployment standard and cluster tests