Added tests for forcing server shutdown.
First I turn DAS into a Zombie - i.e. stop-domain makes DAS undead.  The process keeps running and it is incommunicato.
Then I stop the domain with extreme prejudice via the --force option

The Zombie app (war file) is checked in to svn.  It's not worth the huge effort to make it built as part of the tests.  It's a tiny app and I have checked in the source so changes can easily be made.