Further check-in for secure admin support.
NOTE - enabling secure admin will cause later attempts to submit admin requests to fail.  This is due to some problems in Grizzly that Ryan has already fixed in code and sent to me for testing.  But because of the Grizzly project migration those changes are not yet in the promoted Grizzly that's used for GlassFish builds.

Further, a recently-discovered problem (14381) causes some problems.  Justin or Amy will look into it.

These check-ins embody the asadmin command scripts Justin sent me for enabling and disabling port unification, redirection, and SSL for the admin listener.  

I need to make the config changes more robust and I need to improve message output and error handling, but the normal flow is working.