Fix for 14352
Specifying a non-existent directory resulted in a puzzling error message about a supplemental command failing, with no information about why it had failed.

With this change the DAS "before" supplemental command that checks to make sure a deployed directory is the same on the DAS and on a remote target lets the main deploy command report if the file does not exist.