Fix for 13215
Directory deployment (which does not upload) failed with an IndexOutOfBounds exception when the target was a remote instance.  The DAS had always chosen to upload to a remote instance, but when the only file parameter to the command was a directory there was really nothing to upload.  The code did not take this into account, though, which caused the IOOB exception.  

The fix checks to see if at least one uploadable file parameter was present, and only then actually tries to upload.  
Tests:  QL, deployment devtests, ad hoc tests with a directory deployment to a remote instance