Additional fix for 12989
This change enhances the upgrade to convert the application or module property javaWebStartEnabled to java-web-start-enabled.  (The admin console prior to 3.1 stored the state of the Java Web Start enabled check box using the property name javaWebStartEnabled.  But the code which starts (or not) Java Web Start support for app clients has, since 3.0, looked for java-web-start-enabled.)