Provide a way for the ACC to suppress warnings from the Archivist when both a GlassFish and a legacy Sun runtime descriptor are present in the same archive.
The ACC needs to view the generated app client JAR and the developer's original app client JAR as a single archive.  But doing so can cause the Archivist to find the generated GlassFish runtime DD (in the generated JAR) and the developer's original sun-application-client.xml DD (in the original JAR) from what seems to be the same archive.  If the developer places both DDs into an archive then we want to log a warning, but in this case it is not the developer's fault.

These changes allow the ACC to use an app client archivist that causes the Archivist's method to avoid logging warnings in these cases.

Tests: QL, deployment devtests