Fix for 14934
An earlier change added a config change listener to detect when secure admin was enabled or disabled and to log the need to restart the server.  The listener injected SecureAdmin which is not currently marked as NotNull.  Although this did not seem to bother the normal GlassFish startup it did cause problems with a test in distributions/web and also the embedded QL.  (A subsequent change by Prasad to JMXStartupService has corrected a problem in that code that caused the web distribution test to fail; it did not check for null object names, which apparently the earlier version of this class exposed.)

This change removes the injection of SecureAdmin. Injecting Domain is sufficient to get all relevant notifications.

Tests: product rebuild (distributions/web test) and embedded QL