1. When a persistence bundle has circular dependencies on other bundles which are yet to be resolved or when some dependencies are missing, we are currently not able to enhance, as enhancer tries to resolve and fails. See issue 14998 for details. The new strategy is to defer enhancement to STARTED event if we fail to do it in INSTALLED event. There seems to be a lot of issues while doing it in RESOLVED event, so I decided against it in favour of STARTED event. e.g., when a bundle is getting activated, a bundle may spend a really small time in RESOLVED state, so update() may eventually fail.
2. If we can't enhance in INSTALLED event, we maintain the bundles to be enhanced in a persistent store
so that and the fctually enhance it even if the framework is restarted.

3. I am also adding a capability to configure the enhancer to operate synchronously or asynchronously.
It can be configured by an OSGi bundle context property called org.glassfish.osgijpa.enhancerPolicy.
Valid values: SYNCHRONOUS, ASYNCHRONOUS (both case sensitive)
default value: SYNCHRONOUS
As much as we like to enhance synchronously, there may be potential locking issues
that may prevent enhancement to happen. If that happens, set it to ASYNCHRONOUS.