Fix for 15057
The revised approach for computing the java command to use to start the ACC invokes another small Java program first.  All command-line options and arguments are passed to CLIBootstrap, and as part of that the shell substitutes any environment variables references.  The CLIBootstrap code figures out the java command to be used to launch the ACC.  Any command-line data inside CLIBootstrap has already undergone the shell's env. var. substitution.  When CLIBootstrap emits such data, for non-WIndows systems it now quotes any $ character because any such character is probably part of an unusual but legal path which the user had quoted originally.

(Prior to this change, the un-quoted $ was being interpreted by the shell when the shell processed the java command that invokes the ACC.  We need to suppress that second pass of substitution.)

Tests: QL, deployment and ejb devtests