Fix for 15192
The admin/util module needs the security/core module which contains the SSLUtils class.  The dependency is declared as a "provided" dependency which avoids including the security module and its downstream dependencies in the nucleus distribution.  But the manifest still referred to the package containing SSLUtils, so the nucleus would not start up because the security/core module is (successfully) omitted from the nucleus distribution.

This change makes the resolution of SSLUtils's package optional so it won't be resolved unless used - which it is not in the nucleus distribution.

Change reviewed by/suggested by Sahoo and Snjezana
Approved for check-in by Nazrul

Tests: QL, manual successful start-up of the nucleus distribution