Issue 14992: Adding a fragment to api-exporter bundle. This fragment optionally imports some of the packages that are otherwise imported via DynamicImport-Package in api-exporter bundle. It has been found that statically importing some of the packages sometimes gives us a better result in performance test cases.
Currently, there is only one fragment and it imports all Java EE6 packages (note EE6 also includes SE6).
This fragment is used in all profiles including nucleus. This is not an issue, since it only
imports packages. Ultimately what is visible via this bundle depends on what are the packages exported by
other bundles.

We expect some non-standard packages to be imported by this bundle to boost performance. Santiago is tuning 
the list and is going to add those packages.

Approved for check-in. See JIRA for details.