Fix for GLASSFISH-15267
For @OSGiService whose dynamic=false, the ServiceReference obtained during initialization is checked 
during every method invocation to find if the underlying Service is alive. If the original Service
is not active, a ServiceUnavailableException is thrown, and no attempt is made to rebind to a new Service.

Tests run: QL in Classic and Web profiles, CDI devtests, StockQuote example with dynamic=false
Reviewed by: Sahoo
Approved by: Nazrul