Fix for 13774
On Windows, deploy an EAR then undeploy it. (for example, stateless-session.ear from the devtests) Next, deploy a web app but use the same name as the just-undeployed EAR.  (for example, webapps-caching.war from SQE)  The undeploy of the EAR did not clean out the applications/appName directory because of a locked JAR.  The deployment of the web app fails because GlassFish finds the "left-over" file and tries to treat it as part of the web app - which does not work.

These fixes change the FileArchive implementation so that it recognizes entries only if the entries are dated after a hidden timestamp maintained in the FileArchive's directory.  This hides "left-over" files from a previous deployment that were not cleaned up.

Tests: error scenario in the issue on Windows, QL tests