Fix for http://java.net/jira/browse/GLASSFISH-15881.
The problem was that even though the lazy service initialization was running inside synchronized block, the reference to the service became available before the actual targetInitializer.initializeService() was called. 
So there was a possibility for the newly accepted connection to be passed to the not initialized service, which might lead to unexpected issues depending on timing.

Approved by Chris Kasso.
Patch by Alexey.
Reviewed by Ken and Myself.
Tests run: QL, EJB31