Fix to test for fix to 6545
The test for the fix to 6545 creates a temp directory, populates it, then makes it unreadable so FileArchive cannot list its files.  It seems that on Windows systems the attempt to set the directory unreadable does not work, so during the test FileArchive can read the directory which causes the test to fail (because we expect it not to be able to list the files).

This change checks to see if the test's attempt to make the directory unreadable fails.  If it fails, then the test just returns so it looks successful.