http://java.net/jira/browse/GLASSFISH-16150
Before:
If you gave a relative path (e.g. "./password.txt") for the passwordfile arg, 
this would be stored in the server.  The problem is that the server's cwd is ALWAYS
the config dir.  When you tried to restart it would look for the relative to the config dir --
which was probably wrong.

After:
ProgramOptions.getPasswordFile now "absolutizes" the path automatically.