Fix for 16044
Under cygwin on Windows, a trailing blank on the generated java command line was parsed as an empty additional command-line argument and passed as such to the client. 

This change causes the generated java command to no longer contain an extra trailing blank.

Tests: QL, deployment devtests, cygwin on Windows test