Fix for 16338 in trunk (3.2)
In 3.1 enabling and disabling secure admin required having the cluster-related modules present, because that's where the relevant classes were.

These changes move those classes from cluster/admin to security/core.  Part of the changes was a slight restructuring of how the upgrade class invokes the secure admin logic to avoid having to add a dependency to security/core that would otherwise not be needed.