Fix for 16486
Some com.sun.xxx classes used by Metro are in the extra AIX porting JAR.  This JAR needs to be added to the Class-Path of the gf-client.jar so that the porting JAR is accessible to the Metro JARs which are also specified in the Class-Path.

(The change to PayloadFilesManager brings a fix that was made on the trunk into 3.1.1.  The change suppresses inappropriate warning messages when attempts to create a directory failed because the directory already exists - which is fine and not deserving of a warning.)

Approved for 3.1.1: Sathyan