16762
MiniXmlParser had no idea that the enabled attribute for profiler has a default value of true.  
A unit test immediately broke when I changed it.  Which is nice.  The unit test was also changed to reflect 
the change.

I used the excellent suggestion that Tom added to the issue.  Thanks, Tom!