Incremental fix for 16747
The earlier fix for this performance problem clears the transient app metadata at the end of deployment to release some large data structures.

But redeployment used transient app metadata to let the post deployment command know what targets the app was originally deployed, so the performance fix caused this information to be lost.

These changes store the previous targets in the DeployCommandSupplementalInfo, a separate data structure which conveys some information from the deployment command logic to the PostDeployCommand.  This way, when the transient app metadata is cleared, the post-command still has access to the targets to which the app had been previously deployed.

Tests: QL, deployment cluster devtests