Fix for 17267
GlassFish 3.1.2 sends admin traffic securely between the DAS and the instances, regardless of whether the user has enabled secure admin or not.  This requires some changes in the configuration.  New 3.1.2 domains are created with the correct config.

These changes apply the required changes to a 3.1.1 domain that is being started using 3.1.2.  (Otherwise the DAS and the instances do not communicate with each other correctly so that, for instance, list-instances gives the wrong result or stop-instance fails outright.)

Tests: QL, JIRA issue test case