Additional fix for 17267.
The previous changes handled the "silent upgrade" case when a user starts a 3.1.1 domain using 3.1.2.  This set of changes corrects deficiencies in the formal upgrade logic (that is, when the user upgrades from a 2.x domain to 3.1.2 by running the "asadmin start-domain --upgrade" command.

Basically, this set of changes updates secure admin parts of the Grizzly configuration so that they conform to the 3.1.2 set-up.

Tests: QL, admin cluster-upgrade devtests, scenario in the issue