Fix for 17267 (on the trunk)
GlassFish 3.1.2 and later sends admin traffic securely between the DAS and the instances, regardless of whether the user has enabled secure admin or not.  This requires some changes in the configuration.  New 3.1.2 domains are created with the correct config.

These changes apply the required changes to a 3.1.1 domain that is being started using 3.1.2 or later. (Otherwise the DAS and the instances do not communicate with each other correctly so that, for instance, list-instances gives the wrong result or stop-instance fails outright.)  It also includes changes for the formal upgrade from 2.x to 3.1.2 and later.