Fix for 17371
asadmin can read options (and, if needed someday, other data) from an auxiliary input file using the hidden option _auxinput.  This value is intended to be either a URI or a path (relative or absolute) to a file or "-" indicating that the aux. input should come from stdin.

The logic for differentiating between a URI and a file path was flawed and failed when a valid Windows path included a device letter and the trailing colon.  

These changes repair that bug.

Tests: QL