Changes for 16339
These changes replace sun-acc.xml with glassfish-acc.xml as the default ACC config file, along with parallel changes for replacing sun-application-client-container_1_2.dtd with glassfish-application-client-container_1_3.dtd as the default DTD.  

We keep both sets of files for backward compatibility.  (Users might have scripts which launch app clients that expect sun-acc.xml to be in ${domainDir}/config; if we just replaced sun-acc.xml with glassfish-acc.xml such scripts would break).

These changes are fairly far-reaching.  Changes to the ACC itself use glassfish-acc.xml as the default config file, as do changes in server-side code for preparing Java Web Start launches.  Packaging and domain creation (which now creates both glassfish-acc.xml and sun-acc.xml), etc.  

This check-in also adds the new DTD to the dtds/www directory.

The changes allow users to do an in-place upgrade, which means the new glassfish-acc.xml will not appear in the domain's config directory.  This will be OK, as the CLIBootstrap class looks for sun-acc.xml if it cannot find glassfish-acc.xml.

Tests: QL, deployment devtests