Additional fix for 16302
These changes allow for an in-place update from 3.1.x to 4.0.  If the user does not specify the ACC config file using the -xml option on the appclient command, and if the ACC cannot find the glassfish-acc.xml config file in the domain's directory (it will not be there after an in-place update), then the ACC will look for sun-acc.xml (the old default config file) in the domain's config directory.