Fix for GLASSFISH-17192
Due to a change in the way the lifecycle of Beans is controlled in Weld 1.1.2 (472b728d...), the ThreadContextClassLoader (that held the module ClassLoader of the current BDA being processed) was incorrect in some cases. Some Proxies of Beans in EARs were attempted to be loaded using a module classloader that does not have access to the Bean. Fixed this to ensure that the right ClassLoader (to load the Bean Proxy) is returned for the Bean in ProxyServicesImpl.java

Tests run: QL (with and without Sec Mgr), CDI dev tests including the test referred in GLASSFISH-17192