GLASSFISH-18015
CLI added to start and stop a shared service.
The CLIs are : 1.start-shared-service <servicename> 
	       2. stop-shared-service <servicename>

Start Shared Service: User should be able to start an existing shared service so that applications in the domain can refer to those services.

Stop Shared Service : User should be able to stop a 'running' shared service.
User should not be allowed to stop the shared service that is being referenced by an application that is running.

Reviewed by: Jagadish Ramu

Tests run: QuickLook tests (classic profile and web profile) and PaaS tests