Fix for 18592
MBeans created for config beans use the name as declared in the config bean for part of the MBean name.  For secure-admin-principal elements, the name is the DN which contains characters that are illegal in MBean names.

The constructed ObjectName includes 

name=(name-from-config-bean)

These changes quote the name-from-config-bean if it contains an illegal character.