Various composite resource updates
A new MessageBodyReader, JsonPojoProvider was created to deserialize RestModel instances.  The built-in
Jackson-based support doesn't work as it tries to create a new instance based on the Class of the 
desired type.  Since these are interfaces, the attempt fails, causing the request to error out.

The RestExtension interface was also modified to add support for POST requests.  The details of how this
should work are still being ironed out, so this may very well change, but it's a start.

The composite model generation code was updated to insure that the base interface was added to the list
of interfaces that the generated class implements.

A new utility method, addToParameterMap, was created to assist in transfering data from a RestModel to the
ParameterMap instance used for calling the Set command.  The intent here is to provide a means for building 
the list of Set parameters more simply than requiring the developer to construct the list manually.  It does
so by analzying the target ConfigBean, trying to match its setters against getters on the RestModel.  If there
is a match, the set string is constructed using the basePath provided by the developer.