Fix for 18773.
In certain cases if the DAS does not shut down in response to "stop-domain --force"  the local command logic skipped the code to kill the process because a lower-level exception was thrown but not caught.

These changes catch that exception and, if the user has specified "--kill" uses the OS kill feature.