These changes convert the admin, ReST, and JMX adapters to use the new authentication service for authenticating admin users.
A fair amount of logic has moved from the GenericAdminAuthenticator into the AdminLoginModule and the AdminCallbackHandler.  Further, tokens are associated with the current Subject when the token is created so, when the token is used for later authentication, we can find the saved Subject and establish the correct identity.

Plus, the current Subject is now part of the AdminCommandContext.

pom.xml changes reviewed by Sahoo and Romain
Tests: QL, admin CLI devtests