Fix for 18878.
Use better filtering to decide if a particular MBean method invocation should be subject to access controls when JMX requests are sent to an instance.

The problem in this case was that AMX was being partially started up - during shutdown - just to get its JMX domain name for comparison with the MBean being accessed.  
Tests: QL